+++ packages/common/ast.json
@@ -44,49 +44,6 @@
       "valid": true
     },
     {
-      "name": "fn",
-      "params": [
-        "func"
-      ],
-      "docs": {
-        "description": "Creates a custom step (or operation) for more flexible job writing.",
-        "tags": [
-          {
-            "title": "public",
-            "description": null,
-            "type": null
-          },
-          {
-            "title": "example",
-            "description": "fn(state => {\n  // do some things to state\n  return state;\n});"
-          },
-          {
-            "title": "function",
-            "description": null,
-            "name": null
-          },
-          {
-            "title": "param",
-            "description": "is the function",
-            "type": {
-              "type": "NameExpression",
-              "name": "Function"
-            },
-            "name": "func"
-          },
-          {
-            "title": "returns",
-            "description": null,
-            "type": {
-              "type": "NameExpression",
-              "name": "Operation"
-            }
-          }
-        ]
-      },
-      "valid": true
-    },
-    {
       "name": "jsonValue",
       "params": [
         "obj",
@@ -133,110 +90,6 @@
             "type": {
               "type": "NameExpression",
               "name": "Operation"
-            }
-          }
-        ]
-      },
-      "valid": true
-    },
-    {
-      "name": "sourceValue",
-      "params": [
-        "path"
-      ],
-      "docs": {
-        "description": "Picks out a single value from source data.\nIf a JSONPath returns more than one value for the reference, the first\nitem will be returned.",
-        "tags": [
-          {
-            "title": "public",
-            "description": null,
-            "type": null
-          },
-          {
-            "title": "example",
-            "description": "sourceValue('$.key')"
-          },
-          {
-            "title": "function",
-            "description": null,
-            "name": null
-          },
-          {
-            "title": "param",
-            "description": "JSONPath referencing a point in `state`.",
-            "type": {
-              "type": "NameExpression",
-              "name": "String"
-            },
-            "name": "path"
-          },
-          {
-            "title": "returns",
-            "description": null,
-            "type": {
-              "type": "NameExpression",
-              "name": "Operation"
-            }
-          }
-        ]
-      },
-      "valid": true
-    },
-    {
-      "name": "source",
-      "params": [
-        "path"
-      ],
-      "docs": {
-        "description": "Picks out a value from source data.\nWill return whatever JSONPath returns, which will always be an array.\nIf you need a single value use `sourceValue` instead.",
-        "tags": [
-          {
-            "title": "public",
-            "description": null,
-            "type": null
-          },
-          {
-            "title": "example",
-            "description": "source('$.key')"
-          },
-          {
-            "title": "function",
-            "description": null,
-            "name": null
-          },
-          {
-            "title": "param",
-            "description": "JSONPath referencing a point in `state`.",
-            "type": {
-              "type": "NameExpression",
-              "name": "String"
-            },
-            "name": "path"
-          },
-          {
-            "title": "returns",
-            "description": null,
-            "type": {
-              "type": "TypeApplication",
-              "expression": {
-                "type": "NameExpression",
-                "name": "Array"
-              },
-              "applications": [
-                {
-                  "type": "UnionType",
-                  "elements": [
-                    {
-                      "type": "NameExpression",
-                      "name": "String"
-                    },
-                    {
-                      "type": "NameExpression",
-                      "name": "Object"
-                    }
-                  ]
-                }
-              ]
             }
           }
         ]
@@ -475,59 +328,6 @@
             "type": {
               "type": "NameExpression",
               "name": "array"
-            }
-          }
-        ]
-      },
-      "valid": true
-    },
-    {
-      "name": "each",
-      "params": [
-        "dataSource",
-        "operation"
-      ],
-      "docs": {
-        "description": "Scopes an array of data based on a JSONPath.\nUseful when the source data has `n` items you would like to map to\nan operation.\nThe operation will receive a slice of the data based of each item\nof the JSONPath provided.\n\nIt also ensures the results of an operation make their way back into\nthe state's references.",
-        "tags": [
-          {
-            "title": "public",
-            "description": null,
-            "type": null
-          },
-          {
-            "title": "example",
-            "description": "each(\"$.[*]\",\n  create(\"SObject\",\n    field(\"FirstName\", sourceValue(\"$.firstName\"))\n  )\n)"
-          },
-          {
-            "title": "function",
-            "description": null,
-            "name": null
-          },
-          {
-            "title": "param",
-            "description": "JSONPath referencing a point in `state`.",
-            "type": {
-              "type": "NameExpression",
-              "name": "DataSource"
-            },
-            "name": "dataSource"
-          },
-          {
-            "title": "param",
-            "description": "The operation needed to be repeated.",
-            "type": {
-              "type": "NameExpression",
-              "name": "Operation"
-            },
-            "name": "operation"
-          },
-          {
-            "title": "returns",
-            "description": null,
-            "type": {
-              "type": "NameExpression",
-              "name": "Operation"
             }
           }
         ]
