+++ packages/msgraph/src/Adaptor.js
@@ -53,109 +53,6 @@ export function execute(...operations) {
 }
 
 /**
- * Create some resource in msgraph
- * @public
- * @example
- * create("applications", {"displayName": "My App"})
- * @function
- * @param {string} resource - The type of entity that will be created
- * @param {object} data - The data to create the new resource
- * @param {function} callback - An optional callback function
- * @returns {Operation}
- */
-export function create(resource, data, callback) {
-  return state => {
-    const [resolvedResource, resolvedData] = expandReferences(
-      state,
-      resource,
-      data
-    );
-
-    const { accessToken, apiVersion } = state.configuration;
-
-    const url = setUrl({ apiVersion, resolvedResource });
-
-    const options = {
-      accessToken,
-      body: JSON.stringify(resolvedData),
-      method: 'POST',
-    };
-
-    return request(url, options).then(response =>
-      handleResponse(response, state, callback)
-    );
-  };
-}
-
-/**
- * Make a GET request to msgraph resource
- * @public
- * @example
- *  get('sites/root/lists')
- * @function
- * @param {string} path - Path to resource
- * @param {object} query - Query, Headers and Authentication parameters
- * @param {function} callback - (Optional) Callback function
- * @returns {Operation}
- */
-export function get(path, query, callback = false) {
-  return state => {
-    const { accessToken, apiVersion } = state.configuration;
-    const [resolvedPath, resolvedQuery] = expandReferences(state, path, query);
-
-    const url = setUrl(resolvedPath, apiVersion);
-
-    return request(url, { query: resolvedQuery, accessToken }).then(response =>
-      handleResponse(response, state, callback)
-    );
-  };
-}
-
-/**
- * Get a Drive or SharePoint document library. The drive metadata will be written
- * to state.drives, where it can be used by other adaptor functions.
- * Pass { id } to get a drive by id or { id, owner } to get default drive for
- * some parent resource, like a group
- * @public
- * @example <caption>Get a drive by ID</caption>
- * getDrive({ id: "YXzpkoLwR06bxC8tNdg71m" })
- * @example <caption>Get the default drive for a site</caption>
- * getDrive({ id: "openfn.sharepoint.com", owner: "sites" })
- * @param specifier {Object} - A definition of the drive to retrieve
- *    - id {string} - The ID of the resource or owner.
- *    - owner {string} - The type of drive owner (e.g. sites, groups).
- * @param {string} name - The local name of the drive used to write to state.drives, ie, state.drives[name]
- * @param {function} [callback = s => s] (Optional) Callback function
- * @return {Operation}
- */
-export function getDrive(specifier, name = 'default', callback = s => s) {
-  return state => {
-    const { accessToken, apiVersion } = state.configuration;
-    const [resolvedSpecifier, resolvedName] = expandReferences(
-      state,
-      specifier,
-      name
-    );
-
-    const { id, owner = 'drive' } = resolvedSpecifier;
-
-    let resource;
-    if (owner === 'drive') {
-      resource = `drives/${id}`;
-    } else {
-      resource = `${owner}/${id}/drive`;
-    }
-
-    const url = setUrl(resource, apiVersion);
-
-    return request(url, { accessToken }).then(response => {
-      state.drives[resolvedName] = response;
-      return callback(state);
-    });
-  };
-}
-
-/**
  * Get the contents or metadata of a folder.
  * @public
  * @example <caption>Get a folder by ID</caption>
